+++ app/src/main/java/net/hananet/bns2/poppiyong/PopUpStoreAdpater.kt
@@ -1,4 +1,4 @@
-package com.example.poppiyong_android
+package com.example.ppy_popupstoredetail
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,28 +6,26 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 
-// ToDo: recyclerView의 어댑터
-class HomeBestPopAdapter(private val itemList: List<HomeBestPopModel>) : RecyclerView.Adapter<HomeBestPopAdapter.ViewHolder>() {
-
+/*class PopUpStoreAdpater (private val itemList: List<PopUpStoreModel>) : RecyclerView.Adapter<PopUpStoreAdpater.ViewHolder>(){
     // 뷰 홀더 클래스 정의
     class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-        val popupStoreImageView: ImageView = itemView.findViewById(R.id.homeBestPopImg)
+    //    val popupStoreImageView: ImageView = itemView.findViewById(R.id.popupDetailImg)
     }
 
     // onCreateViewHolder 메서드 오버라이드
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.recyclerview_home_best_item, parent, false)
-        return ViewHolder(view)
-    }
+  //  override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+    //    val view = LayoutInflater.from(parent.context).inflate(R.layout.recyclerview_pop_up_detail_item, parent, false)
+    //    return ViewHolder(view)
+   // }
 
     // onBindViewHolder 메서드 오버라이드
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val item = itemList[position]
-        holder.popupStoreImageView.setImageResource(item.getPopupStoreImg())
+ //       holder.popupStoreImageView.setImageResource(item.getPopupDetailImg())
     }
 
     // getItemCount 메서드 오버라이드
     override fun getItemCount(): Int {
         return itemList.size
     }
-}
+} */
